% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bamboo_meta_fields.R
\name{built_in_bamboo_fields}
\alias{built_in_bamboo_fields}
\title{Aliases and Data Types for built in BambooHR fields}
\usage{
built_in_bamboo_fields()
}
\arguments{
\item{NA}{}
}
\value{
a data frame; containing types and aliases
}
\description{
Return the aliases and data types for BambooHR fields that are not returned
  in the meta data API
}
\details{
BambooHR has a meta data API which lists all discoverable fields along with their types.
  There are additional fields, however, outlined in the
  \href{https://www.bamboohr.com/api/documentation/employees.php}{BambooHR API documentation}
  which are not listed in the meta data API. This function returns those bult in fields.
#'
}
