% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bad_bambooIDs.R
\name{get_bad_bambooIDs}
\alias{get_bad_bambooIDs}
\title{Bamboo API get request wrapper}
\usage{
get_bad_bambooIDs()
}
\value{
tbl_df
}
\description{
Consolidated list of test or incomplete Bamboo IDs
}
\examples{

bamboo_ids <- get_bad_bambooIDs()

}
\references{
\url{https://www.bamboohr.com/api/documentation/},  \url{https://github.com/r-lib/httr}
}
\author{
Evan Downey, \email{edowney@propellerpdx.com}
}
