% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bench.r
\name{get_bench}
\alias{get_bench}
\title{Bamboo API get request wrapper}
\usage{
get_bench(user = NULL, password = NULL, employee_ids = c("all"),
  verbose = FALSE)
}
\arguments{
\item{user}{Bamboo api user id, register in Bamboo "API Keys"}

\item{password}{Bamboo login password}

\item{employee_ids}{an optional list; specifies the employees for which bench time is requested; defaults to c('all') which gets all employee bench time}

\item{verbose}{a logical; indicates if detailed output from httr calls should be provided; default FALSE}
}
\value{
tbl_df
}
\description{
Submits a get request to retrieve the custom bench table for all employees
}
\examples{

user <- 'your_api_user'
password <- 'your_password'
bench <- get_bench(user=user,password=password)

}
\references{
\url{https://www.bamboohr.com/api/documentation/},  \url{https://github.com/r-lib/httr}
}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
