% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_employee_data.R
\name{get_employee_data}
\alias{get_employee_data}
\title{Bamboo API get request wrapper}
\usage{
get_employee_data(ids = NULL, fields = c("firstName", "lastName",
  "jobTitle", "terminationDate", "hireDate", "location"), user = NULL,
  password = NULL, verbose = FALSE)
}
\arguments{
\item{ids}{data frame or list of Bamboo Employee IDs}

\item{fields}{Bamboo fields to be requested, can be requested by id or alias outlined in get_bamboo_meta_fields}

\item{user}{Bamboo api user id, register in Bamboo "API Keys"}

\item{password}{Bamboo login password}

\item{verbose}{a logical; indicates if detailed output from httr calls should be provided; default FALSE}
}
\value{
tbl_df
}
\description{
Submits a get request to retrieve the given fields for the given employees
}
\examples{

ids <- data.frame(id=c(1,6,16,24,44))
fields <- c("firstName","hireDate","location","terminationDate","lastChanged")
user <- 'your_api_user'
password <- 'your_password'
employee_data <- get_employee_data(ids=ids, fields = fields, user=user, password=password)

}
\references{
\url{https://www.bamboohr.com/api/documentation/},  \url{https://github.com/r-lib/httr}
}
\author{
Evan Downey, \email{edowney@propellerpdx.com}
}
