% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_employees_asOf.R
\name{get_employees_asOf}
\alias{get_employees_asOf}
\title{Bamboo API get request wrapper}
\usage{
get_employees_asOf(user = NULL, password = NULL, asOf = Sys.Date(),
  verbose = FALSE)
}
\arguments{
\item{user}{Bamboo api user id, register in Bamboo "API Keys"}

\item{password}{Bamboo login password}

\item{asOf}{Date to get employees}

\item{verbose}{a logical; indicates if detailed output from httr calls should be provided; default FALSE}
}
\value{
tbl_df
}
\description{
Submits a get request to retrieve the all employees as of a past date
}
\examples{

user <- 'your_api_user'
password <- 'your_password'
asOf <- 'date'
employees <- get_employees_asOf(user=user,password=password,asOf=asOf)

}
\references{
\url{https://www.bamboohr.com/api/documentation/},  \url{https://github.com/r-lib/httr}
}
\author{
Evan Downey, \email{edowney@propellerpdx.com}
}
