% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_job_info.R
\name{get_job_info}
\alias{get_job_info}
\title{Bamboo API get request wrapper}
\usage{
get_job_info(user = NULL, password = NULL, verbose = FALSE)
}
\arguments{
\item{user}{Bamboo api user id, register in Bamboo "API Keys"}

\item{password}{Bamboo login password}

\item{verbose}{a logical; indicates if detailed output from httr calls should be provided; default FALSE}
}
\value{
tbl_df
}
\description{
Submits a get request to retrieve the job information history for all employees
}
\examples{

user <- 'your_api_user'
password <- 'your_password'
job_info <- get_job_info(user=user,password=password)

}
\references{
\url{https://www.bamboohr.com/api/documentation/},  \url{https://github.com/r-lib/httr}
}
\author{
Evan Downey, \email{edowney@propellerpdx.com}
}
