% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Bamboo API get request wrapper for BambooHR tables}
\usage{
get_table(requested_table_alias = NULL, user = NULL, password = NULL,
  verbose = FALSE)
}
\arguments{
\item{user}{Bamboo api user id, register in Bamboo "API Keys"}

\item{password}{Bamboo login password}

\item{verbose}{a logical; indicates if detailed output from httr calls should be provided; default FALSE}

\item{table_alias}{the a alis of the table name requested, must be valid from the meta data API}
}
\value{
tbl_df
}
\description{
Submits a get request to retrieve a specified table from the system, valid tables are retreived from the
meta data API
}
\examples{

user <- 'your_api_user'
password <- 'your_password'
verbose <- 'your_preference'
table_alias <- 'table_alias'
table_data <- get_table(table_alias=table_alias, user=user, password=password, verbose=verbose)

}
\references{
\url{https://www.bamboohr.com/api/documentation/},  \url{https://github.com/r-lib/httr}
}
\author{
Evan Downey, \email{edowney@propellerpdx.com}
}
