% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utilization.r
\name{get_utilization}
\alias{get_utilization}
\title{Bamboo API get request wrapper}
\usage{
get_utilization(user = NULL, password = NULL, employee_id = NULL)
}
\arguments{
\item{user}{Bamboo api user id, register in Bamboo "API Keys"}

\item{password}{Bamboo login password}
}
\value{
tbl_df
}
\description{
Submits a get request to retrieve the all employees utilization targets
}
\examples{

user <- 'your_api_user'
password <- 'your_password'
employees <- get_utilization(user=user,password=password,employee_id=c(1,2,3,4))

}
\references{
\url{https://www.bamboohr.com/api/documentation/},  \url{https://github.com/r-lib/httr}
}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
